refactor(opencode): share config dependency scan
